adjust battery value type and clamp to zero when negative

+++ FenomPlus.Mobile/FenomPlus/ViewModels/StatusViewModel.cs
@@ -602,7 +602,7 @@ namespace FenomPlus.ViewModels
             }
             if (Services.DeviceService.Current != null)
             {
-                int value = Services.DeviceService.Current.EnvironmentalInfo.BatteryLevel;
+                float value = Services.DeviceService.Current.EnvironmentalInfo.BatteryLevel;
                 BatteryViewModel.Value = $"{value}%";
                 BatteryViewModel.ButtonText = "Info";
